page-style: Add animated gif for loading
